requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
Django==5.1.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
jiter==0.8.0
numpy==2.1.3
openai==1.56.0
pandas==2.2.3
pydantic==2.10.2
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
tabulate==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
markdown==3.5.2
django-filter==23.5
coreapi==2.3.3
pyyaml==6.0.1

